requirements.txt
pytesseract
numpy
opencv-python
pillow

https://github.com/UB-Mannheim/tesseract/wiki
download tesseractOCR
